requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pytz>=2023.3
typing-extensions>=4.8.0

# Async web framework
aiohttp>=3.9.0

# JSON handling
orjson>=3.9.0

# Logging and monitoring
structlog>=23.2.0

# Redis client (optional, for caching)
redis>=5.0.0

# Security
cryptography>=41.0.0

